if applied, creates the page model and playwright structure test
