+++ routes/caseDetails.ts
@@ -6,7 +6,7 @@ import { handleCaseHistoryTab } from '#src/scripts/controllers/caseHistoryContro
 import { handleCaseDetailsTab, saveProviderNote } from '#src/scripts/controllers/caseDetailsController.js';
 import { getRemoveThirdPartyConfirmation, deleteThirdParty, getRemoveSupportNeedsConfirmation, deleteClientSupportNeeds } from '#src/scripts/controllers/index.js';
 import { getOperatorFeedbackForm, submitOperatorFeedback, getDoYouWantToGiveFeedbackForm, submitDoYouWantToGiveFeedbackForm } from '#src/scripts/controllers/operatorFeedbackController.js';
-import { getSplitThisCaseForm, submitSplitThisCaseForm, getAboutNewCaseForm, submitAboutNewCaseForm } from '#src/scripts/controllers/splitCaseController.js';
+import { getSplitThisCaseForm, submitSplitThisCaseForm, getAboutNewCaseForm, submitAboutNewCaseForm, getCheckSplitCaseAnswersForm } from '#src/scripts/controllers/splitCaseController.js';
 import { validateReopenCase, validateCloseCase, validatePendingCase, validateOperatorFeedback, validateProviderNote, fetchClientDetails, validateGiveFeedback, validateSplitThisCase, validateAboutNewCase } from '#src/middlewares/indexSchema.js';
 
 // Create a new router for case details routes
@@ -146,4 +146,10 @@ router.get('/:caseReference/about-new-case', fetchClientDetails, async (req: Req
 router.post('/:caseReference/about-new-case', fetchClientDetails, validateAboutNewCase(), async (req: Request, res: Response, next: NextFunction): Promise<void> => {
   await submitAboutNewCaseForm(req, res, next);
 });
+
+/* GET check-split-case-answers form. */
+router.get('/:caseReference/check-split-case-answers', fetchClientDetails, async (req: Request, res: Response, next: NextFunction): Promise<void> => {
+  await getCheckSplitCaseAnswersForm(req, res, next);
+});
+
 export default router;
